refactor: faq수정
